Move TSV and HTML output writers into `lirical-io`.
