+++ lirical-core/pom.xml
@@ -48,10 +48,6 @@
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-lang3</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.freemarker</groupId>
-            <artifactId>freemarker</artifactId>
-        </dependency>
 
         <dependency>
             <groupId>org.monarchinitiative.phenol</groupId>
