+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/output/TemplateBasedAnalysisResultsWriter.java
@@ -1,7 +1,11 @@
-package org.monarchinitiative.lirical.core.output;
+package org.monarchinitiative.lirical.io.output;
 
 import org.monarchinitiative.lirical.core.analysis.AnalysisData;
 import org.monarchinitiative.lirical.core.analysis.AnalysisResults;
+import org.monarchinitiative.lirical.core.output.AnalysisResultsMetadata;
+import org.monarchinitiative.lirical.core.output.AnalysisResultsWriter;
+import org.monarchinitiative.lirical.core.output.OutputFormat;
+import org.monarchinitiative.lirical.core.output.OutputOptions;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
 import org.monarchinitiative.phenol.ontology.data.Ontology;
