+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/output/svg/Posttest2Svg.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.lirical.core.output.svg;
+package org.monarchinitiative.lirical.io.output.svg;
 
 import org.monarchinitiative.lirical.core.analysis.AnalysisResults;
 import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDisease;
