+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/output/svg/package-info.java
@@ -1,4 +1,4 @@
 /** Classes to create an SVG graphic representing the result of likelihood raio analysis of one HPO Case.
  * @author <a href="mailto:peter.robinson@jax.org">Peter Robinson</a>
  */
-package org.monarchinitiative.lirical.core.output.svg;
+package org.monarchinitiative.lirical.io.output.svg;
