+++ lirical-io/src/main/resources/org/monarchinitiative/lirical/io/output/liricalTSV.ftl
