+++ lirical-io/src/test/java/org/monarchinitiative/lirical/io/output/JsonAnalysisResultWriterTest.java
@@ -27,7 +27,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
